+++ serverside/makoto/src/test/kotlin/FeatureDemonstration.kt
@@ -227,7 +227,7 @@ val FeatureDemonstration by testSuite {
                             teamIdentifier = "9CYHJNG644",
                             bundleIdentifier = "at.asitplus.attestation-client",
                             sandbox = false,
-                            trustedRootOverrides = setOf(APPLE_DEFAULT_TRUSTED_ROOTS.receiptRoot to APPLE_DEFAULT_TRUSTED_ROOTS.receiptRoot)
+                            trustedRootOverrides = setOf(APPLE_DEFAULT_TRUSTED_ROOTS.first().receiptRoot to APPLE_DEFAULT_TRUSTED_ROOTS.first().receiptRoot)
                         )
                     ),
                     iosVersion = IosAttestationConfiguration.OsVersions(